Fixes https://app.clickup.com/t/9002019994/PRF-4737

+++ apps/manager_app/manager_api.h
@@ -56,5 +56,5 @@ bool decode_manager_query(const uint8_t *data,
 bool encode_manager_result(manager_result_t *result,
                            uint8_t *buffer,
                            uint16_t max_buffer_len,
-                           uint32_t *bytes_written_out);
+                           size_t *bytes_written_out);
 #endif
